+++ Charts/Classes/Components/ChartLimitLine.swift
@@ -27,7 +27,7 @@ public class ChartLimitLine: ChartComponentBase
     }
     
     /// limit / maximum (the y-value or xIndex)
-    public var limit = Float(0.0)
+    public var limit = Double(0.0)
     
     private var _lineWidth = CGFloat(2.0)
     public var lineColor = UIColor(red: 237.0/255.0, green: 91.0/255.0, blue: 91.0/255.0, alpha: 1.0)
@@ -43,20 +43,20 @@ public class ChartLimitLine: ChartComponentBase
         super.init();
     }
     
-    public init(limit: Float)
+    public init(limit: Double)
     {
         super.init();
         self.limit = limit;
     }
     
-    public init(limit: Float, label: String)
+    public init(limit: Double, label: String)
     {
         super.init();
         self.limit = limit;
         self.label = label;
     }
     
-    /// set the line width of the chart (min = 0.2f, max = 12f); default 2f
+    /// set the line width of the chart (min = 0.2, max = 12); default 2
     public var lineWidth: CGFloat
     {
         get
